refactored for routes
